+++ Application/UseCases/Products/BaseCommands/ProductHardDelete.cs
@@ -1,10 +1,10 @@
-using Application.UseCases.BaseServices.HardDelete;
-using Domain.Entities.ProductModule.ProductAggregate;
+using Application.UseCases.Base.HardDelete;
+using Domain.ProductModule.Entities;
 using SharedKernel.Interfaces;
 
 namespace Application.UseCases.Products.BaseCommands
 {
-    public record ProductHardDeleteCommand(List<int> Ids, int UserId) : GenericHardDeleteCommand(Ids, UserId);
+    public record ProductHardDeleteCommand(List<long> Ids, long UserId) : GenericHardDeleteCommand(Ids, UserId);
 
     public class ProductHardDeleteHandler : GenericHardDeleteHandler<Product, ProductHardDeleteCommand>
     {
