+++ Application/UseCases/Products/BaseCommands/ProductRecovery.cs
@@ -1,10 +1,10 @@
-using Application.UseCases.BaseServices.Recovery;
-using Domain.Entities.ProductModule.ProductAggregate;
+using Application.UseCases.Base.Recovery;
+using Domain.ProductModule.Entities;
 using SharedKernel.Interfaces;
 
 namespace Application.UseCases.Products.BaseCommands
 {
-    public record ProductRecoveryCommand(int Id, int UserId) : GenericRecoveryCommand(Id, UserId);
+    public record ProductRecoveryCommand(long Id, long UserId) : GenericRecoveryCommand(Id, UserId);
 
     public class ProductRecoveryHandler : GenericRecoveryHandler<Product, ProductRecoveryCommand>
     {
